fix: 트렌딩 Redis score 경계 처리 보강

+++ trending-service/src/main/java/com/sungho/letterpick/trending/ranking/application/PublicIssueRankingScoreCalculator.java
@@ -8,9 +8,9 @@ import java.util.Objects;
 @Component
 public class PublicIssueRankingScoreCalculator {
 
-    public int calculate(RedisPublicIssueRankingStateReader.AvailableIssueRankingState state) {
+    public long calculate(RedisPublicIssueRankingStateReader.AvailableIssueRankingState state) {
         Objects.requireNonNull(state, "state must not be null");
 
-        return Math.toIntExact(state.viewCount());
+        return state.viewCount();
     }
 }
